Updated Setting screen structure
