+++ sample-app/App.tsx
@@ -7,7 +7,7 @@ import {
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 import Home from './screens/Home/Home'
-import Settings from './screens/Settings'
+import Settings from './screens/Settings/Settings'
 
 const Stack = createStackNavigator();
