+++ sample-app/helpers/AsyncStore.tsx
@@ -1,0 +1,6 @@
+import { createStore } from 'redux'
+
+function emptyHook() { }
+
+export let store = createStore(emptyHook)
+export const storageKey = "storage_key"
